requirements.txt
langsmith
google-adk
langchain-google-genai~=4.1.3
python-dotenv~=1.2.1
langchain-community~=0.4.1
tavily-python
fastapi~=0.123.10
uvicorn
pydantic~=2.12.5
langchain~=1.2.3
langgraph~=1.0.5
typing_extensions~=4.15.0